+++ localDB.json
@@ -1,9 +1,1 @@
-{
-  "onerm": {
-    "bench press": 0,
-    "squat": 0,
-    "deadlift": 0,
-    "overhead press": 0,
-    "bent over row": 0
-  }
-}
+{"onerm":{"bench press":"100","squat":0,"deadlift":0,"overhead press":0,"bent over row":0}}
